+++ includes/class-activitypub.php
@@ -148,7 +148,7 @@ class Activitypub {
 				 */
 				$activitypub_template = apply_filters( 'activitypub_preview_template', ACTIVITYPUB_PLUGIN_DIR . '/templates/post-preview.php' );
 			} else {
-				$activitypub_template = ACTIVITYPUB_PLUGIN_DIR . '/templates/activitypub-json.php';
+				$activitypub_template = ACTIVITYPUB_PLUGIN_DIR . 'templates/activitypub-json.php';
 			}
 		}
 
@@ -169,10 +169,17 @@ class Activitypub {
 		}
 
 		if ( $activitypub_template ) {
+			\set_query_var( 'is_404', false );
+
 			// Check if header already sent.
-			if ( ! \headers_sent() && ACTIVITYPUB_SEND_VARY_HEADER ) {
-				// Send Vary header for Accept header.
-				\header( 'Vary: Accept' );
+			if ( ! \headers_sent() ) {
+				// Send 200 status header.
+				\status_header( 200 );
+
+				if ( ACTIVITYPUB_SEND_VARY_HEADER ) {
+					// Send Vary header for Accept header.
+					\header( 'Vary: Accept' );
+				}
 			}
 
 			return $activitypub_template;