+++ CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Enabled querying of Outbox posts through the REST API to improve troubleshooting and debugging.
 * Updated terminology to be client-neutral in the Federated Reply block.
 
+### Fixed
+
+* Enforce 200 status header for valid ActivityPub requests.
+
 ## [5.1.0] - 2025-02-06
 
 ### Added